pyproject.toml
[project]
name = "frango"
version = "0.1.0"
description = ""
authors = [ ]
requires-python = "<4.0,>=3.12"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "loguru>=0.7.2",
    "dataclass-wizard>=0.22.2",
    "rraft-py>=0.2.27",
    "grpcio>=1.62.2",
    "sqlglot>=23.12.1",
]
devDependencies = [
    "grpcio-tools>=1.64.0",
]

[project.scripts]
frango-node = "frango.node:main"
frango-cli = "frango.client.cli:main"

[tool.pdm.scripts]
node = { call = "frango.node:main" }
cli = { call = "frango.client.cli:main" }

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = [ "frango/" ]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

